+++ gens/models/genomic.py
@@ -124,7 +124,7 @@ class GenomicRegion(RWModel):
             raise ValueError(f"{start} is not a valid start position")
         return region
 
-    @computed_field()
+    @computed_field() # type: ignore
     @property
     def chromosome(self) -> Chromosome:
         """Get the chromosome name from region string."""
@@ -134,7 +134,7 @@ class GenomicRegion(RWModel):
             raise ValueError("Invalid region designation.")
         return Chromosome(match.group(1))
 
-    @computed_field()
+    @computed_field() # type: ignore
     @property
     def start(self) -> int | None:
         """Get start position from a region string."""
@@ -142,7 +142,7 @@ class GenomicRegion(RWModel):
         match = re.match(REGION_PATTERN, self.region)
         return int(match.group(2)) if match else None
 
-    @computed_field()
+    @computed_field() # type: ignore
     @property
     def end(self) -> int | None:
         """Get end position from a region string."""