+++ gens/utils.py
@@ -5,4 +5,4 @@ import datetime
 
 def get_timestamp() -> datetime.datetime:
     """Get datetime timestamp in utc timezone."""
-    return datetime.datetime.now(tz=datetime.UTC)
+    return datetime.datetime.now(tz=datetime.timezone.utc)