+++ CHANGELOG.md
@@ -22,6 +22,7 @@ About changelog [here](https://keepachangelog.com/en/1.0.0/)
  - Initial type hints added to command and db modules
  - Remove unused file argument for load chromomsome info CLI command
  - Fixed various issues raised by pylint
+ - mypy type fixing for `mypy gens/commands/load.py`
 
 ## 3.0.1